Se completa el formateo de campos del excel ©kristian-hernandez

+++ muric-entities/src/main/java/co/com/muric/entities/util/InformacionCredito.java
@@ -18,7 +18,7 @@ public class InformacionCredito {
     private Integer calidadDeudor;
     private Integer fechaDesembolso;
     private Integer fechaVencimiento;
-    private Double valorDesembolsado;
+    private Float valorDesembolsado;
     private Integer frecuenciaPagoCapital;
     private Integer frecuenciaPagoIntereses;
     private String tipoTasa;
